Keep expected behavior for dav unlink and rmdir

+++ src/plugins/http/gfal_http_plugin_metadata.cpp
@@ -37,6 +37,22 @@ int gfal_http_mkdirpG(plugin_handle plugin_data, const char* url, mode_t mode, g
 int gfal_http_unlinkG(plugin_handle plugin_data, const char* url, GError** err){
     GfalHttpInternal* davix = gfal_http_get_plugin_context(plugin_data);
     Davix::DavixError* daverr = NULL;
+
+    // Note: in WebDAV DELETE works for directories and files, so check ourselves
+    // we are not unlinking a folder
+    struct stat st;
+    if (davix->posix.stat(&davix->params, url, &st, &daverr) != 0) {
+      davix2gliberr(daverr, err);
+      Davix::DavixError::clearError(&daverr);
+      return -1;
+    }
+
+    if (S_ISDIR(st.st_mode)) {
+        gfal2_set_error(err, http_plugin_domain, EISDIR, __func__,
+                  "Can not unlink a directory");
+        return -1;
+    }
+
     if (davix->posix.unlink(&davix->params, url, &daverr) != 0) {
       davix2gliberr(daverr, err);
       Davix::DavixError::clearError(&daverr);
@@ -50,6 +66,22 @@ int gfal_http_unlinkG(plugin_handle plugin_data, const char* url, GError** err){
 int gfal_http_rmdirG(plugin_handle plugin_data, const char* url, GError** err){
     GfalHttpInternal* davix = gfal_http_get_plugin_context(plugin_data);
     Davix::DavixError* daverr = NULL;
+
+    // Note: in WebDAV DELETE is recursive for directories, so check first if there is anything
+    // inside and fail in that case
+    struct stat st;
+    if (davix->posix.stat(&davix->params, url, &st, &daverr) != 0) {
+      davix2gliberr(daverr, err);
+      Davix::DavixError::clearError(&daverr);
+      return -1;
+    }
+
+    if (!S_ISDIR(st.st_mode)) {
+        gfal2_set_error(err, http_plugin_domain, ENOTDIR, __func__,
+                  "Can not rmdir a file");
+        return -1;
+    }
+
     if (davix->posix.rmdir(&davix->params, url, &daverr) != 0) {
       davix2gliberr(daverr, err);
       Davix::DavixError::clearError(&daverr);
